Cast result_size to size_t at CONCAT ER_QPROC_STRING_SIZE_TOO_BIG sites
qstr_concatenate and qstr_bit_concatenate emit message 1042 through the
error_status variable, so they were missed when the message switched to
%zu. *result_size is int; under %zu the error manager reads 8 bytes for
it and corrupts the second placeholder. Cast both to (size_t).

+++ src/query/string_opfunc.c
@@ -9170,7 +9170,7 @@ qstr_concatenate (const unsigned char *s1, int s1_length, int s1_size_, int s1_p
 
 size_error:
   error_status = ER_QPROC_STRING_SIZE_TOO_BIG;
-  er_set (ER_ERROR_SEVERITY, ARG_FILE_LINE, error_status, 2, *result_size,
+  er_set (ER_ERROR_SEVERITY, ARG_FILE_LINE, error_status, 2, (size_t) * result_size,
 	  (int) prm_get_bigint_value (PRM_ID_STRING_MAX_SIZE_BYTES));
   return error_status;
   /*
@@ -9382,7 +9382,7 @@ qstr_bit_concatenate (const unsigned char *s1, int s1_length, int s1_precision, 
 
 size_error:
   error_status = ER_QPROC_STRING_SIZE_TOO_BIG;
-  er_set (ER_ERROR_SEVERITY, ARG_FILE_LINE, error_status, 2, *result_size,
+  er_set (ER_ERROR_SEVERITY, ARG_FILE_LINE, error_status, 2, (size_t) * result_size,
 	  (int) prm_get_bigint_value (PRM_ID_STRING_MAX_SIZE_BYTES));
   return error_status;
